slimmed orchestration
